# crates/puddle-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "puddle",
    "description": "puddle's local API: pending requests, rules, audit, settings, consents and the event stream. Listens on 127.0.0.1 only. Every request needs `Authorization: Bearer <token>` (from the connection file), a `Host` of `127.0.0.1:<port>` or `localhost:<port>`, and no foreign `Origin`.",
    "contact": {
      "name": "Tijs van Kampen",
      "email": "puddle@tijsvankampen.be"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Every workspace.",
        "operationId": "list_workspaces",
        "responses": {
          "200": {
            "description": "the workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceList"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Creates a workspace: its disk, then a clone of the repository. Answers 202; follow it with\n`workspace_progress` events.",
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "creation started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "the name is taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "invalid name, repository URL (including SSH remotes), image or memory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "workspaces are not available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "One workspace.",
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "Deletes a stopped workspace, its disk included. The body must confirm. The service checks\nagain first and refuses (409) unless the fresh check has the `fingerprint` the user saw, or\nis clean when none is sent. Answers 202.",
        "operationId": "delete_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "delete accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "running, busy, or changed since it was checked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "not confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/attach": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Opens a running workspace in VS Code: on the desktop (puddle opens it) or in the browser\n(the response has the URL, the caller opens it).",
        "operationId": "attach_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "what was done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachResponse"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "not running, or direct SSH is off for a desktop attach",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/delete-check": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "What deleting the workspace would lose: uncommitted changes, unpushed commits, stashes and\ndata outside any checkout. Can take a while (it looks inside the workspace).",
        "operationId": "delete_check",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCheck"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/reclaim": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Gives the disk space the workspace freed back to the host. Answers 202.",
        "operationId": "reclaim_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "reclaim accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "busy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/start": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Starts a workspace that is down. Answers 202.",
        "operationId": "start_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "workspace id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "start accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "no such workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "already running or busy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, lower case, no trailing period. Never contains a token."
          }
        }
      },
      "AttachMode": {
        "type": "string",
        "description": "How to open the workspace in an editor.",
        "enum": [
          "desktop",
          "browser"
        ]
      },
      "AttachRequest": {
        "type": "object",
        "description": "`POST /api/workspaces/{id}/attach`.",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AttachMode",
            "description": "Where to open it."
          }
        },
        "additionalProperties": false
      },
...
      "DeleteWorkspaceRequest": {
        "type": "object",
        "description": "`DELETE /api/workspaces/{id}`.",
        "required": [
          "confirm"
        ],
        "properties": {
          "confirm": {
            "type": "boolean",
            "description": "Must be `true`: the user confirmed."
          },
          "fingerprint": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `fingerprint` of the delete check the user saw. Left out or `null` means \"I saw\nnothing to lose\": the delete goes ahead only if a fresh check is clean."
          }
        },
        "additionalProperties": false
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "What went wrong, as a stable code a client can switch on.",
        "enum": [
          "misdirected_host",
          "forbidden_origin",
          "unauthorized",
          "bad_request",
          "unsupported_media_type",
          "payload_too_large",
          "invalid",
          "not_found",
          "conflict",
          "newer_settings",
          "unavailable",
          "internal"
        ]
      },
...
      "NewWorkspaceRequest": {
        "type": "object",
        "description": "`POST /api/workspaces`.",
        "required": [
          "name",
          "repo_url"
        ],
        "properties": {
          "image": {
            "type": [
              "string",
              "null"
            ],
            "description": "The image to boot; left out or `null` is the default devcontainer image."
          },
          "memory_mib": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Memory in MiB (256 to 1048576); left out or `null` is the default.",
            "maximum": 1048576,
            "minimum": 256
          },
          "name": {
            "$ref": "#/components/schemas/WorkspaceName",
            "description": "The workspace's name, which is also its workspace's name."
          },
          "repo_url": {
            "type": "string",
            "description": "The repository to clone, as an `https://` URL without credentials. SSH remotes\n(`git@host:path`, `ssh://`) are refused with a message saying so."
          }
        },
        "additionalProperties": false
      },
...
      "WorkspaceList": {
        "type": "object",
        "description": "`GET /api/workspaces`.",
        "required": [
          "workspaces"
        ],
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            },
            "description": "Every workspace, ordered by id."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}